requirements.txt
Flask
streamlit
pandas
python-dotenv
requests
